Allow naming the Git remote.
By default, Git remotes are named 'origin', but some users of rpkg might
be interacting with several remotes for the same package module.

This change allows sites to define how remotes should be called by
setting the appropriate variable in the configuration file.

If nothing is specified, the default is 'origin', to not break existing
clones.

This has been tested with both regular clones and clones with folders
for each branch.

The rpkg bash completion resource had to be modified accoringly. Since
the fedpkg one already handles this properly, a simple copy-paste of
that specific line was enough.

Nothing else is assuming the remote to be called 'origin', so everything
should go on working as expected.